refactor: move default impls from iroh-bytes into iroh and make them private
since we are using Arc<dyn ...> the type of the dummy handler does not need
to be public.
