+++ iroh/Cargo.toml
@@ -16,6 +16,7 @@ rust-version = "1.65"
 anyhow = { version = "1", features = ["backtrace"] }
 bao-tree = { version = "0.5.0", features = ["tokio_fsm"], default-features = false }
 blake3 = "1.3.3"
+bytes = "1"
 derive_more = { package = "derive_more_preview", version = "0.1.0", features = ["debug", "display", "from", "try_into"] }
 flume = "0.10.14"
 futures = "0.3.25"
